# swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: Dawarich API
  version: v1
  description: API for Dawarich, your favorite alternative to Google Timeline. Provides
    endpoints for managing location data, tracks, visits, statistics, and more.
paths:
...
  "/api/v1/settings":
    patch:
      summary: Updates user settings
      tags:
      - Settings
      parameters:
      - name: api_key
        in: query
        required: true
        description: API Key
        schema:
          type: string
      responses:
        '200':
          description: settings updated
        '401':
          description: unauthorized
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                route_opacity:
                  type: number
                  example: 60
                  description: Route opacity percentage (0-100)
                meters_between_routes:
                  type: number
                  example: 500
                  description: Minimum distance between routes in meters
                minutes_between_routes:
                  type: number
                  example: 30
                  description: Minimum time between routes in minutes
                time_threshold_minutes:
                  type: number
                  example: 30
                  description: Time threshold for grouping points in minutes
                merge_threshold_minutes:
                  type: number
                  example: 15
                  description: Threshold for merging nearby points in minutes
                preferred_map_layer:
                  type: string
                  example: OpenStreetMap
                  description: Preferred map layer/tile provider
                speed_colored_routes:
                  type: boolean
                  example: false
                  description: Whether to color routes based on speed
                points_rendering_mode:
                  type: string
                  example: raw
                  description: How to render points on the map (raw, heatmap, etc.)
                live_map_enabled:
                  type: boolean
                  example: true
                  description: Whether live map updates are enabled
                immich_url:
                  type: string
                  example: https://immich.example.com
                  description: Immich server URL for photo integration
                immich_api_key:
                  type: string
                  example: your-immich-api-key
                  description: API key for Immich photo service
                photoprism_url:
                  type: string
                  example: https://photoprism.example.com
                  description: PhotoPrism server URL for photo integration
                photoprism_api_key:
                  type: string
                  example: your-photoprism-api-key
                  description: API key for PhotoPrism photo service
                speed_color_scale:
                  type: string
                  example: viridis
                  description: Color scale for speed-colored routes
                maps:
                  type: object
                  properties:
                    distance_unit:
                      type: string
                      enum:
                      - km
                      - mi
                      example: km
                      description: Distance unit used for displayed distances
            examples:
              '0':
                summary: Updates user settings
                value:
                  settings:
                    route_opacity: 60
                    meters_between_routes: 500
                    minutes_between_routes: 30
                    time_threshold_minutes: 30
                    merge_threshold_minutes: 15
                    preferred_map_layer: OpenStreetMap
                    speed_colored_routes: false
                    points_rendering_mode: raw
                    live_map_enabled: true
                    immich_url: https://immich.example.com
                    immich_api_key: your-immich-api-key
                    photoprism_url: https://photoprism.example.com
                    photoprism_api_key: your-photoprism-api-key
                    speed_color_scale: viridis
                    maps:
                      distance_unit: km
    get:
      summary: Retrieves user settings
      tags:
      - Settings
      parameters:
      - name: api_key
        in: query
        required: true
        description: API Key
        schema:
          type: string
      responses:
        '200':
          description: settings found
          content:
            application/json:
              schema:
                type: object
                properties:
                  settings:
                    type: object
                    properties:
                      route_opacity:
                        type: number
                        example: 60
                        description: Route opacity percentage (0-100)
                      meters_between_routes:
                        oneOf:
                        - type: number
                        - type: string
                        example: 500
                        description: Minimum distance between routes in meters
                      minutes_between_routes:
                        oneOf:
                        - type: number
                        - type: string
                        example: 30
                        description: Minimum time between routes in minutes
                      time_threshold_minutes:
                        oneOf:
                        - type: number
                        - type: string
                        example: 30
                        description: Time threshold for grouping points in minutes
                      merge_threshold_minutes:
                        oneOf:
                        - type: number
                        - type: string
                        example: 15
                        description: Threshold for merging nearby points in minutes
                      preferred_map_layer:
                        type: string
                        example: OpenStreetMap
                        description: Preferred map layer/tile provider
                      speed_colored_routes:
                        type: boolean
                        example: false
                        description: Whether to color routes based on speed
                      points_rendering_mode:
                        type: string
                        example: raw
                        description: How to render points on the map (raw, heatmap,
                          etc.)
                      live_map_enabled:
                        type: boolean
                        example: true
                        description: Whether live map updates are enabled
                      immich_url:
                        oneOf:
                        - type: string
                        - type: 'null'
                        example: https://immich.example.com
                        description: Immich server URL for photo integration
                      immich_api_key:
                        oneOf:
                        - type: string
                        - type: 'null'
                        example: your-immich-api-key
                        description: API key for Immich photo service
                      photoprism_url:
                        oneOf:
                        - type: string
                        - type: 'null'
                        example: https://photoprism.example.com
                        description: PhotoPrism server URL for photo integration
                      photoprism_api_key:
                        oneOf:
                        - type: string
                        - type: 'null'
                        example: your-photoprism-api-key
                        description: API key for PhotoPrism photo service
                      speed_color_scale:
                        oneOf:
                        - type: string
                        - type: 'null'
                        example: viridis
                        description: Color scale for speed-colored routes
        '401':
          description: unauthorized
...
servers:
- url: http://{defaultHost}
  variables:
    defaultHost:
      default: localhost:3000
components:
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: query
      description: API key passed as a query parameter
    bearer_auth:
      type: http
      scheme: bearer
      description: Bearer token authentication